V2.1 prerelease

+++ codegen/src/main/java/com/oes/openfmb/generation/proto/Helpers.java
@@ -41,7 +41,7 @@ class Helpers {
                     Int32Value.getDescriptor(),
                     Int64Value.getDescriptor(),
                     UInt32Value.getDescriptor(),
-                    UInt32Value.getDescriptor()
+                    UInt64Value.getDescriptor()
             )
     );
